scripts/requirements.txt
psycopg2-binary==2.9.9
openai==2.8.1
python-dotenv==1.0.0